Reduce engine dependencies. crtlib/mathlib/matrixlib now can be built standalone
