+++ ref_gl/gl_backend.c
@@ -16,7 +16,6 @@ GNU General Public License for more details.
 
 #include "gl_local.h"
 #include "mathlib.h"
-#include "common.h"
 
 char		r_speeds_msg[MAX_SYSPATH];
 ref_speeds_t	r_stats;	// r_speeds counters
