+++ ref_gl/gl_decals.c
@@ -15,7 +15,6 @@ GNU General Public License for more details.
 
 #include "gl_local.h"
 #include "cl_tent.h"
-#include "common.h"
 
 #define DECAL_OVERLAP_DISTANCE	2
 #define DECAL_DISTANCE		4	// too big values produce more clipped polygons
