+++ ref_gl/gl_refrag.c
@@ -16,7 +16,6 @@ GNU General Public License for more details.
 #include "gl_local.h"
 #include "entity_types.h"
 #include "studio.h"
-#include "common.h"
 #include "world.h" // BOX_ON_PLANE_SIDE
 
 /*
